requirements.txt
Babel==2.8.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
filelock==3.13.3
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
forex-python==1.8
greenlet==3.0.3
httplib2==0.20.2
idna==3.6
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.3
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.1.5
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
platformdirs==4.2.0
psycopg2-binary==2.9.9
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu3
pytz==2022.1
PyYAML==5.4.1
requests==2.31.0
SecretStorage==3.3.1
simplejson==3.19.2
six==1.16.0
SQLAlchemy==2.0.29
systemd-python==234
typing_extensions==4.11.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.1
virtualenv==20.25.1
wadllib==1.3.6
Werkzeug==3.0.2
zipp==1.0.0
